Move model command into own file
