+++ cli.js
@@ -43,59 +43,6 @@ try {
   console.warn(err);
 } */
 
-const _modelApp = async output => {
-  const bundleBuffer = fs.readFileSync(output);
-  const bundle = new wbn.Bundle(bundleBuffer);
-
-  const res = bundle.getResponse('https://xrpackage.org/manifest.json');
-  const s = res.body.toString('utf8');
-  const manifestJson = JSON.parse(s);
-  const {start_url: startUrl, xr_type: xrType} = manifestJson;
-
-  const builder = cloneBundle(bundle, {
-    except: ['/manifest.json'],
-  });
-
-  let modelIcon = manifestJson.icons && manifestJson.icons.find(icon => icon.type === 'model/gltf-binary');
-  if (!modelIcon) {
-    let modelPath;
-    switch (xrType) {
-      case 'gltf@0.0.1':
-      case 'vrm@0.0.1':
-      {
-        /* const res = bundle.getResponse(primaryUrl + '/' + startUrl);
-        return res.body; */
-        modelPath = startUrl;
-        break;
-      }
-      default: {
-        modelPath = 'xrpackage_model.glb';
-
-        const modelUint8Array = fs.readFileSync(path.join(__dirname, 'assets', 'w.glb'));
-        builder.addExchange(primaryUrl + '/' + modelPath, 200, {
-          'Content-Type': 'model/gltf-binary',
-        }, modelUint8Array);
-        break;
-      }
-    }
-
-    modelIcon = {
-      src: modelPath,
-      type: 'model/gltf-binary',
-    };
-    if (!Array.isArray(manifestJson.icons)) {
-      manifestJson.icons = [];
-    }
-    manifestJson.icons.push(modelIcon);
-  }
-
-  builder.addExchange(primaryUrl + '/manifest.json', 200, {
-    'Content-Type': 'application/json',
-  }, JSON.stringify(manifestJson, null, 2));
-
-  const buffer = builder.createBundle();
-  fs.writeFileSync(output, buffer);
-};
 const _bakeApp = async output => {
   const app = express();
   app.use((req, res, next) => {
@@ -262,20 +209,7 @@ yargs
   .command(require('./commands/build'))
   .command(require('./commands/screenshot'))
   .command(require('./commands/volume'))
-  .command('model [input]', 'generate a model of the package at [input]', yargs => {
-    yargs
-      .positional('input', {
-        describe: 'built package to model (a.wbn)',
-      });
-  }, async argv => {
-    handled = true;
-
-    if (typeof argv.input !== 'string') {
-      argv.input = 'a.wbn';
-    }
-
-    await _modelApp(argv.input);
-  })
+  .command(require('./commands/model'))
   .command('bake [input]', 'bake screenshot/volume/model of the package at [input]', yargs => {
     yargs
       .positional('input', {
